Separando lógica de pré processamento do HeaderSmv
